requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gspread==6.1.2
google-auth==2.29.0
pydantic==2.7.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openpyxl>=3.1.2
